Better fix for timer + minor progression
song metadata + progress updates every 20sec as it was at 1sec before. That was too fast for the .json file from azuracast API refresh's that happens in at least ~15s as i observed, It will be changed later with better measurements from the uptade time of API.
early beginning of cover displaying, started write a few lines of code about, it not working yet.